Completed JWT Auth Backend

+++ src/main/java/com/example/chamudika/repositories/UserRepository.java
@@ -1,10 +1,9 @@
 package com.example.chamudika.repositories;
+
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import com.example.chamudika.models.User;
-
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> { 
     Optional<User> findByEmail(String email);
 }
